+++ ergotree-interpreter/src/eval/option_is_defined.rs
@@ -1,0 +1,52 @@
+use ergotree_ir::mir::option_is_defined::OptionIsDefined;
+use ergotree_ir::mir::value::Value;
+
+use crate::eval::env::Env;
+use crate::eval::EvalContext;
+use crate::eval::EvalError;
+use crate::eval::Evaluable;
+
+impl Evaluable for OptionIsDefined {
+    fn eval(&self, env: &Env, ctx: &mut EvalContext) -> Result<Value, EvalError> {
+        let v = self.input.eval(env, ctx)?;
+        match v {
+            Value::Opt(opt_v) => Ok(opt_v.is_some().into()),
+            _ => Err(EvalError::UnexpectedExpr(format!(
+                "Don't know how to eval OptM: {0:?}",
+                self
+            ))),
+        }
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::OptionIsDefined;
+    use crate::eval::context::Context;
+    use crate::eval::tests::eval_out;
+    use ergotree_ir::mir::expr::Expr;
+    use ergotree_ir::mir::extract_reg_as::ExtractRegisterAs;
+    use ergotree_ir::mir::global_vars::GlobalVars;
+    use ergotree_ir::types::stype::SType;
+    use sigma_test_util::force_any_val;
+    use std::rc::Rc;
+
+    #[test]
+    fn eval() {
+        let get_reg_expr: Expr = ExtractRegisterAs::new(
+            GlobalVars::SelfBox.into(),
+            0,
+            SType::SOption(SType::SLong.into()),
+        )
+        .unwrap()
+        .into();
+        let option_expr: Expr = OptionIsDefined {
+            input: Box::new(get_reg_expr),
+        }
+        .into();
+        let ctx = Rc::new(force_any_val::<Context>());
+        let v = eval_out::<bool>(&option_expr, ctx);
+        // R0 is always defined (box value)
+        assert_eq!(v, true);
+    }
+}